Move bert to model_schedule
